# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WRS Digital Agricultural Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING_SETTLEMENT, SETTLED, EXPIRED, CANCELLED]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: executeOrder
      tags: [orders]
      summary: Execute a buy order against a spot listing (Off-Taker only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created and eWR locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request or listing not available
        "403":
          description: Forbidden
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status (confirm settlement or cancel)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "403":
          description: Forbidden
        "404":
          description: Not found
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, name, email, tier, reputationScore, kybStatus, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        tier:
          type: string
          enum: [PRODUCER, OFF_TAKER, ENABLER, FINANCIER]
        reputationScore:
          type: integer
        kybStatus:
          type: string
          enum: [PENDING, VERIFIED, REJECTED]
        company:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        company:
          type: string
        tier:
          type: string
          enum: [PRODUCER, OFF_TAKER, ENABLER, FINANCIER]
...
    Order:
      type: object
      required: [id, listingId, buyerId, lockedAt, expiresAt, status, totalUsd, platformFeeUsd, escrowFeeUsd, createdAt]
      properties:
        id:
          type: integer
        listingId:
          type: integer
        buyerId:
          type: integer
        buyerName:
          type: ["string", "null"]
        lockedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        settledAt:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [PENDING_SETTLEMENT, SETTLED, EXPIRED, CANCELLED]
        totalUsd:
          type: number
        platformFeeUsd:
          type: number
        escrowFeeUsd:
          type: number
        commodityType:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        weightMt:
          type: ["number", "null"]
        warehouseCode:
          type: ["string", "null"]
        pricePerMt:
          type: ["number", "null"]
        sellerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    OrderInput:
      type: object
      required: [listingId]
      properties:
        listingId:
          type: integer
    OrderUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [SETTLED, CANCELLED]
...