# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WRS Digital Agricultural Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  listingId: number;
  buyerId: number;
  /** @nullable */
  buyerName?: string | null;
  lockedAt: Date;
  expiresAt: Date;
  /** @nullable */
  settledAt?: Date | null;
  status: OrderStatus;
  totalUsd: number;
  platformFeeUsd: number;
  escrowFeeUsd: number;
  /** @nullable */
  commodityType?: string | null;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  weightMt?: number | null;
  /** @nullable */
  warehouseCode?: string | null;
  /** @nullable */
  pricePerMt?: number | null;
  /** @nullable */
  sellerName?: string | null;
  createdAt: Date;
}