# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WRS Digital Agricultural Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { UserKybStatus } from './userKybStatus';
import type { UserTier } from './userTier';

export interface User {
  id: number;
  clerkId: string;
  name: string;
  email: string;
  tier: UserTier;
  reputationScore: number;
  kybStatus: UserKybStatus;
  /** @nullable */
  company?: string | null;
  createdAt: Date;
}