# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WRS Digital Agricultural Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateTier } from './userUpdateTier';

export interface UserUpdate {
  name?: string;
  company?: string;
  tier?: UserUpdateTier;
}